Unified account-menu.
