+++ webofneeds/won-owner-webapp/src/main/webapp/app/components/account-menu.js
@@ -1,0 +1,104 @@
+/**
+ * Created by ksinger on 01.09.2017.
+ */
+;
+import angular from 'angular';
+import { attach } from '../utils';
+import { actionCreators }  from '../actions/actions';
+import {
+    connect2Redux,
+} from '../won-utils';
+
+import dropdownModule from './covering-dropdown';
+import loginFormModule from './login-form';
+
+import * as srefUtils from '../sref-utils';
+
+function genLogoutConf() {
+    let template = `
+        <won-dropdown>
+            <won-dd-header class="topnav__button">
+
+
+                <span class="topnav__button__caption">
+                    {{self.loggedIn? self.email : "Sign In"}}
+                </span>
+                <img
+                    src="generated/icon-sprite.svg#ico16_arrow_down"
+                    class="topnav__carret">
+                <img
+                    class="topnav__button__icon"
+                    src="generated/icon-sprite.svg#ico36_person">
+
+
+            </won-dd-header>
+            <won-dd-menu>
+
+
+                <div class="am__menu--loggedin" ng-show="self.loggedIn">
+                    <a
+                        href="{{ self.absHRef(self.$state, 'about') }}"
+                        class="topnav__button red">
+                            About
+                    </a>
+                    <button
+                        class="won-button--filled lighterblue"
+                        style="width:100%"
+                        ng-click="::self.logout()">
+                            Sign out
+                    </button>
+                </div>
+
+
+                <won-login-form
+                    class="am__menu--loggedout"
+                    ng-show="!self.loggedIn">
+                </won-login-form>
+
+
+            </won-dd-menu>
+        </won-dropdown>
+    `;
+
+    const serviceDependencies = ['$state', '$ngRedux', '$scope', /*'$routeParams' /*injections as strings here*/];
+
+    class Controller {
+        constructor(/* arguments <- serviceDependencies */){
+            attach(this, serviceDependencies, arguments);
+            Object.assign(this, srefUtils);
+
+            window.lopu4dbg = this;
+
+            this.email = "";
+            this.password = "";
+
+            const logout = (state) => ({
+                loggedIn: state.getIn(['user','loggedIn']),
+                email: state.getIn(['user','email'])
+            });
+
+            connect2Redux(logout, actionCreators, [], this);
+        }
+
+        hideLogin() {
+            this.open = false;
+        }
+    }
+    Controller.$inject = serviceDependencies;
+
+    return {
+        restrict: 'E',
+        controller: Controller,
+        controllerAs: 'self',
+        bindToController: true, //scope-bindings -> ctrl
+        scope: {open: '='},
+        template: template
+    }
+}
+
+export default angular.module('won.owner.components.accountMenu', [
+    dropdownModule,
+    loginFormModule,
+])
+    .directive('wonAccountMenu', genLogoutConf)
+    .name;
