+++ webofneeds/won-owner-webapp/src/main/webapp/style/won.scss
@@ -144,3 +144,4 @@ hr {
 @import 'send-request';
 @import 'post-owner';
 @import 'connections-map';
+@import 'covering-dropdown';
